+++ Crow/Crow.csproj
@@ -4,13 +4,14 @@
 		<TargetFramework>netstandard2.0</TargetFramework>
 		
 		<AssemblyVersion>$(CrowVersion)</AssemblyVersion>
-		<ReleaseVersion>0.8.0</ReleaseVersion>
+		<ReleaseVersion>$(CrowVersion)</ReleaseVersion>
+		<PackageVersion>$(CrowPackageVersion)</PackageVersion>
+		
 		<Title>C# Rapid Open Widget Toolkit</Title>
 		<Description>C.R.O.W. is a widget toolkit and rendering engine developed in C# with templates, styles, compositing, and bindings.</Description>
 		<License>MIT</License>
 		<Authors>Jean-Philippe Bruyère</Authors>
 		<RepositoryUrl>https://github.com/jpbruyere/Crow</RepositoryUrl>
-		<PackageVersion>$(CrowPackageVersion)</PackageVersion>
 		<PackageTags>GUI Widget toolkit Interface C# .Net Mono</PackageTags>
 		<PackageRequireLicenseAcceptance>False</PackageRequireLicenseAcceptance>
 		<PackageProjectUrl>https://github.com/jpbruyere/Crow/wiki</PackageProjectUrl>
@@ -38,7 +39,7 @@
 		<PackageReference Include="System.Reflection.Emit.ILGeneration" Version="4.6.0" />
 		<PackageReference Include="System.Reflection.Emit.Lightweight" Version="4.6.0" />
 		<!--<PackageReference Include="glfw-sharp" Version="0.2.0" />-->
-		<PackageReference Include="glfw-sharp" Version="0.2.2" />
+		<PackageReference Include="glfw-sharp" Version="0.2.4" />
 	</ItemGroup>
 	<PropertyGroup Condition=" '$(CrowStbSharp)' == 'true'">
 		<DefineConstants>$(DefineConstants);STB_SHARP</DefineConstants>
