+++ Crow/src/EventArgs/MouseCursorChangedEventArgs.cs
